Rename mod `mcsv` -> `stdin_sender`
