+++ src/server/stdin_sender.rs
